# resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Dataspace Ecosystem APIs
  description: Merged OpenAPI specification for all Dataspace Ecosystem extensions
  license:
    name: "The Apache License, Version 2.0"
    url: http://www.apache.org/licenses/LICENSE-2.0.txt
  version: 0.14.0
servers:
- url: /
security:
- bearerAuth: []
- apiKeyAuth: []
...
paths:
  /billing-reports:
    get:
      tags:
      - Telemetry CSV Manager API
      description: Retrieves Reports
      operationId: getReport
      parameters:
      - name: month
        in: query
        description: Target month
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: year
        in: query
        description: Target year
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: The telemetry event was processed successfully
          content:
            text/csv:
              schema:
                type: string
        "400":
          description: Invalid date range provided
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Invalid JWT token
          content:
            application/json:
              schema:
                type: string
        "403":
          description: "Missing/invalid participant in roles, unexpected number of\
            \ participant roles or participant does not exist"
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
...
components:
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      name: x-api-key
      in: header
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT